+++ CHANGELOG.md
@@ -35,6 +35,7 @@ All notable changes to the Wazuh app project will be documented in this file.
 - Changed `rule.groups` filter to `wazuh.integration.decoders` [#7934](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7934)
 - Applied the new home page navigation style to all dashboards [#7981](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7981)
 - Updated Office 365 dashboards to use new index pattern [#8081](https://github.com/wazuh/wazuh-dashboard-plugins/pull/8081)
+- Reintroduce Google Cloud dashboard [#8069](https://github.com/wazuh/wazuh-dashboard-plugins/pull/8069)
 
 ### Fixed
 
@@ -77,7 +78,7 @@ All notable changes to the Wazuh app project will be documented in this file.
 - Fixed `indexDate` function to correctly handle ISO week-year calculation for weekly indices [#8047](https://github.com/wazuh/wazuh-dashboard-plugins/pull/8047)
 - Fixed the status code and body displayed in failed requests in Server management > Dev tools as internal server error [#8052](https://github.com/wazuh/wazuh-dashboard-plugins/pull/8052)
 
-## Wazuh v4.14.3 - OpenSearch Dashboards 2.19.4 - Revision 01
+## Wazuh v4.14.3 - OpenSearch Dashboards 2.19.4 - Revision 02
 
 ### Added